+++ gluten-substrait/src/main/scala/org/apache/gluten/extension/columnar/EnsureLocalSortRequirements.scala
@@ -16,9 +16,7 @@
  */
 package org.apache.gluten.extension.columnar
 
-import org.apache.gluten.GlutenConfig
-import org.apache.gluten.extension.columnar.MiscColumnarRules.TransformPreOverrides
-import org.apache.gluten.extension.columnar.rewrite.RewriteSparkPlanRulesManager
+import org.apache.gluten.extension.columnar.heuristic.HeuristicTransform
 
 import org.apache.spark.sql.catalyst.expressions.SortOrder
 import org.apache.spark.sql.catalyst.rules.Rule
@@ -33,24 +31,14 @@ import org.apache.spark.sql.execution.{SortExec, SparkPlan}
  * SortAggregate with the same key. So, this rule adds local sort back if necessary.
  */
 object EnsureLocalSortRequirements extends Rule[SparkPlan] {
-  private lazy val offload = TransformPreOverrides.apply()
+  private lazy val transform = HeuristicTransform()
 
   private def addLocalSort(
       originalChild: SparkPlan,
       requiredOrdering: Seq[SortOrder]): SparkPlan = {
+    // FIXME: HeuristicTransform is costly. Re-applying it may cause performance issues.
     val newChild = SortExec(requiredOrdering, global = false, child = originalChild)
-    if (!GlutenConfig.getConf.enableColumnarSort) {
-      FallbackTags.add(newChild, "columnar Sort is not enabled in SortExec")
-      newChild
-    } else {
-      val rewrittenPlan = RewriteSparkPlanRulesManager.apply().apply(newChild)
-      if (rewrittenPlan.eq(newChild) && FallbackTags.nonEmpty(rewrittenPlan)) {
-        // The sort can not be offloaded
-        rewrittenPlan
-      } else {
-        offload.apply(rewrittenPlan)
-      }
-    }
+    transform.apply(newChild)
   }
 
   override def apply(plan: SparkPlan): SparkPlan = {
