+++ gluten-substrait/src/main/scala/org/apache/gluten/extension/columnar/heuristic/LegacyOffload.scala
@@ -1,0 +1,48 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.gluten.extension.columnar.heuristic
+
+import org.apache.gluten.logging.LogLevelUtil
+
+import org.apache.spark.sql.catalyst.rules.Rule
+import org.apache.spark.sql.execution.SparkPlan
+
+class LegacyOffload(topDownRules: Seq[OffloadSingleNode], bottomUpRules: Seq[OffloadSingleNode])
+  extends Rule[SparkPlan]
+  with LogLevelUtil {
+
+  def apply(plan: SparkPlan): SparkPlan = {
+    val plan0 =
+      topDownRules.foldLeft(plan)((p, rule) => p.transformDown { case p => rule.offload(p) })
+    val plan1 =
+      bottomUpRules.foldLeft(plan0)((p, rule) => p.transformUp { case p => rule.offload(p) })
+    plan1
+  }
+}
+
+object LegacyOffload {
+  def apply(): LegacyOffload = {
+    new LegacyOffload(
+      List(),
+      List(
+        OffloadOthers(),
+        OffloadExchange(),
+        OffloadJoin()
+      )
+    )
+  }
+}
